[Feat] 이미지처리 서비스 코드 추가 #8
