--- src/main/java/com/example/medicinebackend/OpenAPI/FeignController.java
@@ -1,19 +1,0 @@
-package com.example.medicinebackend.OpenAPI;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class FeignController {
-    private final OpenFeignService openFeignService;
-
-    @GetMapping("/search")
-    public Object getMedicineDataByName(
-            @RequestParam("productName") String productName) {
-        return openFeignService.getMedicineDataByName(productName);
-    }
-
-}
